// binders-devops-service-v1/app/src/dc/docker-compose-spec.d.ts
export type ListOfStrings = string[];
// ...
export interface Service {
// ...
    depends_on?:
        | ListOfStrings
        | {
              /**
               * This interface was referenced by `undefined`'s JSON-Schema definition
               * via the `patternProperty` "^[a-zA-Z0-9._-]+$".
               */
              [k: string]: {
                  restart?: boolean | string;
                  required?: boolean;
                  condition: "service_started" | "service_healthy" | "service_completed_successfully";
                  /**
                   * This interface was referenced by `undefined`'s JSON-Schema definition
                   * via the `patternProperty` "^x-".
                   */
                  [k: string]: unknown;
              };
          };
// ...
}